# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ADEX documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "ADEX API Server"
        }
    ],
    "paths": {
...
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Получить личную информацию пользователя",
                "operationId": "getUserInfo",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "firstname": {
                                            "type": "string",
                                            "example": "Ivan"
                                        },
                                        "lastname": {
                                            "type": "string",
                                            "example": "Ivan"
                                        },
                                        "about": {
                                            "type": "string",
                                            "example": "Я пользователь биржы ADEX"
                                        },
                                        "phone": {
                                            "type": "integer",
                                            "example": "79549765434"
                                        },
                                        "city": {
                                            "type": "string",
                                            "example": "Москва"
                                        },
                                        "country": {
                                            "type": "string",
                                            "example": "Россия"
                                        },
                                        "rating": {
                                            "type": "string",
                                            "example": "Россия"
                                        },
                                        "employment_type": {
                                            "type": "string",
                                            "example": "Россия"
                                        },
                                        "avatar": {
                                            "type": "string",
                                            "example": "https::/127.0.0.1/storage/123.png"
                                        },
                                        "success_reviews": {
                                            "type": "string",
                                            "example": "5"
                                        },
                                        "failed_reviews": {
                                            "type": "string",
                                            "example": "2"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2021-04-04"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Изменить личную информацию пользователя",
                "operationId": "updateUserInfo",
                "requestBody": {
                    "description": "Заполните поля для изменения личной информации пользователя (employment_type: 1 - свободен, 2 - занят)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "Ivan"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Ivan"
                                    },
                                    "about": {
                                        "type": "string",
                                        "example": "Я пользователь биржы ADEX"
                                    },
                                    "phone": {
                                        "type": "integer",
                                        "example": "79549765434"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Москва"
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Россия"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "21-04-1980"
                                    },
                                    "employment_type": {
                                        "type": "ште",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Удаление своей учётной записи",
                "operationId": "deleteUser",
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/user/avatar": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Изменить аватар",
                "operationId": "changeAvatar",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "image": {
                                            "type": "string",
                                            "example": "https::/127.0.0.1/storage/123.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed change user avatar",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed change user avatar"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    }
}